Liquid Glass Icon
Did this over in STE as well. Was curious about what this might look like for this app. Made the 'theme' color, of sorts, a bit brighter, as the contrast with the blue piece wasn't showing very nicely. May reflect this to the app theme color itself eventually, as well. Will see.